+++ lib/construct/type.ex
@@ -38,6 +38,14 @@ defmodule Construct.Type do
 
   ## Functions
 
+  def builtin do
+    @builtin
+  end
+
+  def builtinc do
+    @builtinc
+  end
+
   @doc """
   Checks if we have a primitive type.
 
@@ -104,14 +112,14 @@ defmodule Construct.Type do
       iex> cast(:binary, "beef")
       {:ok, "beef"}
 
-      iex> cast(:decimal, Decimal.new(1.0))
-      {:ok, Decimal.new(1.0)}
+      iex> cast(:decimal, Decimal.from_float(1.0))
+      {:ok, Decimal.from_float(1.0)}
       iex> cast(:decimal, Decimal.new("1.0"))
-      {:ok, Decimal.new(1.0)}
+      {:ok, Decimal.from_float(1.0)}
       iex> cast(:decimal, 1.0)
-      {:ok, Decimal.new(1.0)}
+      {:ok, Decimal.from_float(1.0)}
       iex> cast(:decimal, "1.0")
-      {:ok, Decimal.new(1.0)}
+      {:ok, Decimal.from_float(1.0)}
 
       iex> cast({:array, :integer}, [1, 2, 3])
       {:ok, [1, 2, 3]}
@@ -146,6 +154,7 @@ defmodule Construct.Type do
         else
           type.cast(term)
         end
+
       true ->
         cast(type, term)
     end
@@ -205,8 +214,7 @@ defmodule Construct.Type do
   def cast(:reference, term) when is_reference(term), do: {:ok, term}
 
   def cast(:decimal, term) when is_binary(term) do
-    apply(Decimal, :parse, [term])
-    |> validate_decimal()
+    validate_decimal(apply(Decimal, :parse, [term]))
   end
   def cast(:decimal, term) when is_integer(term) do
     {:ok, apply(Decimal, :new, [term])}
@@ -245,8 +253,10 @@ defmodule Construct.Type do
     cond do
       not primitive?(type) ->
         type.cast(term)
+
       of_base_type?(type, term) ->
         {:ok, term}
+
       true ->
         :error
     end
@@ -382,178 +392,6 @@ defmodule Construct.Type do
       :error ->
         :error
     end
-  end
-
-  ## Typespecs
-
-  @doc """
-  Returns typespec AST for given type
-
-    iex> spec([CommaList, {:array, :integer}]) |> Macro.to_string()
-    "list(:integer)"
-
-    iex> spec({:array, :string}) |> Macro.to_string()
-    "list(String.t())"
-
-    iex> spec({:map, CustomType}) |> Macro.to_string()
-    "%{optional(term) => CustomType.t()}"
-
-    iex> spec(:string) |> Macro.to_string()
-    "String.t()"
-
-    iex> spec(CustomType) |> Macro.to_string()
-    "CustomType.t()"
-  """
-  @spec spec(t) :: Macro.t()
-
-  def spec(type) when is_list(type) do
-    type |> List.last() |> spec()
-  end
-
-  def spec({:array, type}) do
-    quote do
-      list(unquote(spec(type)))
-    end
-  end
-
-  def spec({:map, type}) do
-    quote do
-      %{optional(term) => unquote(spec(type))}
-    end
-  end
-
-  def spec({typec, _arg}) do
-    quote do
-      unquote(typec).t()
-    end
-  end
-
-  def spec(:string) do
-    quote do
-      String.t()
-    end
-  end
-
-  def spec(:decimal) do
-    quote do
-      Decimal.t()
-    end
-  end
-
-  def spec(:utc_datetime) do
-    quote do
-      DateTime.t()
-    end
-  end
-
-  def spec(:naive_datetime) do
-    quote do
-      NaiveDateTime.t()
-    end
-  end
-
-  def spec(:date) do
-    quote do
-      Date.t()
-    end
-  end
-
-  def spec(:time) do
-    quote do
-      Time.t()
-    end
-  end
-
-  def spec(type) when type in @builtin do
-    type
-  end
-
-  def spec(type) when is_atom(type) do
-    quote do
-      unquote(type).t()
-    end
-  end
-
-  def spec(type) do
-    type
-  end
-
-  @doc """
-  Returns typespec AST for given term
-
-    iex> typeof(nil) |> Macro.to_string()
-    "nil"
-
-    iex> typeof(1.42) |> Macro.to_string()
-    "float()"
-
-    iex> typeof("string") |> Macro.to_string()
-    "String.t()"
-
-    iex> typeof(CustomType) |> Macro.to_string()
-    "CustomType.t()"
-
-    iex> typeof(&NaiveDateTime.utc_now/0) |> Macro.to_string()
-    "NaiveDateTime.t()"
-  """
-  @spec typeof(t) :: Macro.t()
-
-  def typeof(term) when is_nil(term) do
-    nil
-  end
-
-  def typeof(term) when is_integer(term) do
-    {:integer, [], []}
-  end
-
-  def typeof(term) when is_float(term) do
-    {:float, [], []}
-  end
-
-  def typeof(term) when is_boolean(term) do
-    {:boolean, [], []}
-  end
-
-  def typeof(term) when is_binary(term) do
-    quote do
-      String.t()
-    end
-  end
-
-  def typeof(term) when is_pid(term) do
-    {:pid, [], []}
-  end
-
-  def typeof(term) when is_reference(term) do
-    {:reference, [], []}
-  end
-
-  def typeof(%{__struct__: struct}) when is_atom(struct) do
-    quote do
-      unquote(struct).t()
-    end
-  end
-
-  def typeof(term) when is_map(term) do
-    {:map, [], []}
-  end
-
-  def typeof(term) when is_atom(term) do
-    quote do
-      unquote(term).t()
-    end
-  end
-
-  def typeof(term) when is_list(term) do
-    {:list, [], []}
-  end
-
-  def typeof(term) when is_function(term, 0) do
-    term.() |> typeof()
-  end
-
-  def typeof(_) do
-    {:term, [], []}
   end
 
   ## Helpers