+++ test/support/test_case.ex
@@ -19,15 +19,31 @@ defmodule Construct.TestCase do
     end
   end
 
+  defmacro create_module([do: block]) do
+    create_test_module(block)
+  end
+
   defmacro create_construct([do: block]) do
-    __create_construct__([], block)
+    create_test_module(quote do
+      use Construct
+
+      structure do
+        unquote(block)
+      end
+    end)
   end
 
   defmacro create_construct(opts, [do: block]) do
-    __create_construct__(opts, block)
+    create_test_module(quote do
+      use Construct, unquote(opts)
+
+      structure do
+        unquote(block)
+      end
+    end)
   end
 
-  def __create_construct__(opts, block) do
+  defp create_test_module(ast) do
     quote do
       # retrieve module name from test case and line
       name = __ENV__.module
@@ -36,11 +52,7 @@ defmodule Construct.TestCase do
       module_name = :"#{name}_#{line}"
 
       defmodule module_name do
-        use Construct, unquote(opts)
-
-        structure do
-          unquote(block)
-        end
+        unquote(ast)
       end
     end
   end