+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## v3.0.0
 
-<!-- TODO fill the changelog -->
+* Enhancements
+  * Add `structure_compile_hook/2` macro to inject code into structures
+    See `Construct.Hooks.Map` and `Construct.Hooks.OmitDefault` for examples
+  * Compiler optimization and refactor
+
+* Hard-deprecations
+  * Remove `make_map` option
+  * Remove `empty_values` option
 
 ## v2.1.10